requirements.txt
streamlit
audio-recorder-streamlit
openai
tqdm
python-dotenv
watchdog
whisperplus
transformers
pyannote.audio>=3.1.0
pyannote.core>=5.0.0
pyannote.database>=5.0.1
pyannote.metrics>=3.2.1
pyannote.pipeline>=3.0.1
torchaudio>=2.0.0
speechbrain
huggingface_hub[cli]